requirements.txt
# Web Framework
fastapi==0.104.1
pydantic==2.5.0

# ASGI Server
uvicorn[standard]==0.24.0

# HTTP Client
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0